Hint at future parts and note the past, e.g. Aristotle
Note early on, in Chapter 4, that later sections of the book
will expand TFL to create FOL and then modal logic.
I think it helps students to have brief notes, before they dive in,
of what will lay ahead.

In addition, briefly note Aristotle, Frege, and Russell.
Aristotle is mentioned later, but not *why*. Many students
won't know that Aristotle created logic as an area, and may not
even know who Aristotle is (!). Make sure that a reader
*knows*  that Aristotle got the study of logic started.
Also mention that the logic systems described
in this book are modern logic, which is related to but not the same as
Aristotelian logic. Also briefly mention Frege and Russell.
While this isn't a history book, it's important for students
to learn that this material didn't just get provided by magic,
but that instead there were *people* behind its development
and a *history* of its development.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>