Material connditional is also called material implication
Clarify early on that this definition of the conditional is
the *material* conditional (readers might otherwise not realize this),
and make it clear that an alternative term is "material implication".
The term material implication is also widely used, and it's important
that students see common terms so that they can begin to find the relevant
literature.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>