A small correction in the answer to question 4 of exercise A in chapter 28.
The question says:
"Only Ingmar knows the combination to the safe"
(this answer as the comment below says, is taking true the case where Ingmar doesn't know the combination, so I respect that and my correction is not about that detail)

The answer written in the solutionary is: "∀x[ K(i) → ( x=i ) ]" 
This isn't correct, because it says, that for all the elements in the domain, if "K(i)", then all the elements are "i". Which is only true if in the domain is just composed by the element "i", but false if not.

What I guess the author was referring to, is to "∀x[ K(x) → ( x=i ) ]", that, all the elements that fulfill "K" are "i". So, no element that isn't "i" is "K".

So, I just replace the letter "i" with the letter "x" :D